+++ src/main/java/me/matsubara/roulette/util/InventoryUpdate.java
@@ -21,7 +21,6 @@
  */
 package me.matsubara.roulette.util;
 
-import com.cryptomorin.xseries.ReflectionUtils;
 import org.apache.commons.lang.Validate;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
