+++ src/main/java/me/matsubara/roulette/util/ReflectionUtils.java
@@ -19,7 +19,7 @@
  * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
  * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.cryptomorin.xseries;
+package me.matsubara.roulette.util;
 
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
